Moved some classes that should not be publicly visible.
