+++ src/main/java/org/jsapar/TextComposer.java
@@ -1,9 +1,9 @@
 package org.jsapar;
 
 import org.jsapar.compose.Composer;
-import org.jsapar.compose.ComposerFactory;
-import org.jsapar.compose.SchemaComposer;
-import org.jsapar.compose.TextComposerFactory;
+import org.jsapar.compose.internal.ComposerFactory;
+import org.jsapar.compose.internal.SchemaComposer;
+import org.jsapar.compose.internal.TextComposerFactory;
 import org.jsapar.error.ErrorEventListener;
 import org.jsapar.model.Document;
 import org.jsapar.model.Line;
