+++ src/main/java/org/jsapar/compose/bean/BeanComposer.java
@@ -2,7 +2,7 @@ package org.jsapar.compose.bean;
 
 import org.jsapar.compose.ComposeException;
 import org.jsapar.compose.Composer;
-import org.jsapar.compose.ValidationHandler;
+import org.jsapar.compose.line.ValidationHandler;
 import org.jsapar.error.ErrorEvent;
 import org.jsapar.error.ErrorEventListener;
 import org.jsapar.error.ExceptionErrorEventListener;
