+++ src/main/java/org/jsapar/compose/fixed/FixedWidthComposer.java
@@ -1,7 +1,7 @@
 package org.jsapar.compose.fixed;
 
-import org.jsapar.compose.AbstractSchemaComposer;
-import org.jsapar.compose.SchemaComposer;
+import org.jsapar.compose.internal.AbstractSchemaComposer;
+import org.jsapar.compose.internal.SchemaComposer;
 import org.jsapar.schema.FixedWidthSchema;
 import org.jsapar.schema.FixedWidthSchemaLine;
