+++ src/main/java/org/jsapar/compose/internal/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Internal composer classes and interfaces.
+ */
+package org.jsapar.compose.internal;
