+++ src/main/java/org/jsapar/text/TextParseConfig.java
@@ -4,8 +4,6 @@ import org.jsapar.error.ValidationAction;
 
 /**
  * Configuration that controls behavior while parsing text.
- * <p>
- * Created by stejon0 on 2016-07-12.
  */
 public class TextParseConfig {
